+++ config/settings.py
@@ -26,6 +26,7 @@ INSTALLED_APPS = [
     "rest_framework",
     "rest_framework_simplejwt",
     "users",
+    "payments",
     "course",
     "django_filters",
 ]
