+++ config/urls.py
@@ -5,4 +5,5 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("course/", include("course.urls", namespace="course")),
     path("user/", include("users.urls", namespace="user")),
+    path("payments/", include("payments.urls", namespace="payments")),
 ]
