+++ payments/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework.serializers import ModelSerializer
+
+from payments.models import Payments
+
+class PaymentsSerializer(ModelSerializer):
+
+    class Meta:
+        model = Payments
+        fields = "__all__"
