+++ users/admin.py
@@ -1,3 +1,8 @@
 from django.contrib import admin
 
-# Register your models here.
+from users.models import User
+
+
+@admin.register(User)
+class UserAdmin(admin.ModelAdmin):
+    list_display = ('id', 'email')
