+++ users/serializers.py
@@ -1,12 +1,13 @@
 from rest_framework.fields import SerializerMethodField
 from rest_framework.serializers import ModelSerializer
-from users.models import Payments, User
+
+from payments.models import Payments
+from users.models import User
 
 
 class UserSerializer(ModelSerializer):
 
     payments_history = SerializerMethodField()
-
 
     def get_payments_history(self, user):
         return Payments.objects.filter(user=user).values(
@@ -26,8 +27,4 @@ class UserSerializer(ModelSerializer):
         )
 
 
-class PaymentsSerializer(ModelSerializer):
 
-    class Meta:
-        model = Payments
-        fields = "__all__"
