+++ users/views.py
@@ -1,12 +1,9 @@
-from django.shortcuts import render
-from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework import filters
 from rest_framework.generics import CreateAPIView
 from rest_framework.permissions import AllowAny
 from rest_framework.viewsets import ModelViewSet
 
-from users.models import Payments, User
-from users.serializers import PaymentsSerializer, UserSerializer
+from users.models import User
+from users.serializers import UserSerializer
 
 
 class UserCreateApiView(CreateAPIView):
@@ -20,23 +17,8 @@ class UserCreateApiView(CreateAPIView):
         user.save()
 
 
-
 class UserViewSet(ModelViewSet):
     queryset = User.objects.all()
     serializer_class = UserSerializer
 
 
-class PaymentsViewSet(ModelViewSet):
-    queryset = Payments.objects.all()
-    filter_backends = [DjangoFilterBackend, filters.OrderingFilter]
-    filterset_fields = [
-        "paid_course",
-        "paid_lesson",
-        "payment_method",
-    ]
-    ordering_fields = [
-        "payment_date",
-        "payment_amount",
-    ]
-
-    serializer_class = PaymentsSerializer
